Moved Metadata enums to enums package
