+++ core/src/main/java/io/github/scroojalix/npcmanager/commands/subcommands/modifications/meta/GlowColorMeta.java
@@ -10,7 +10,7 @@ import org.bukkit.command.CommandSender;
 import io.github.scroojalix.npcmanager.NPCMain;
 import io.github.scroojalix.npcmanager.commands.SubCommand;
 import io.github.scroojalix.npcmanager.npc.NPCData;
-import io.github.scroojalix.npcmanager.npc.meta.MetaColor;
+import io.github.scroojalix.npcmanager.npc.meta.enums.MetaColor;
 import io.github.scroojalix.npcmanager.utils.PluginUtils;
 
 public class GlowColorMeta extends SubCommand {
